Fix my busted prompt
The prompt was having a problem where if you up-arrowed thru bash
history it would start printing garbage.

The problem was that bash was having trouble trying to determine
the length of the prompt because of special characters used to
colorize it.  By surrounding these things in \[...\] bracers, it
helps bash compute the length better.  now history works great!